requirements.txt
Flask==2.0.2
python-dotenv==0.19.1
pymysql==1.0.2
